# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Digital Twin API",
    "description": "Digital Twin - Car Health Monitoring Platform backend. Phase 1 provides vehicle registration and raw telemetry storage.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a user account",
        "description": "Register a new account. The caller never supplies a role.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Register Api V1 Auth Register Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Exchange credentials for a token pair",
        "description": "Login. A 401 is returned for a bad email, a bad password, and an\ninactive account alike, so the existence of an email can never be probed.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Rotate a refresh token for a fresh token pair",
        "description": "Each refresh rotates the token; the previous one is revoked.",
        "operationId": "refresh_api_v1_auth_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Revoke a refresh token",
        "description": "Revoke the given refresh token. Idempotent: repeating a logout or\nlogging out with an unknown token still returns 204.",
        "operationId": "logout_api_v1_auth_logout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get the authenticated user's profile",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/vehicles/{vehicle_id}/agent/query": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Ask a natural-language question about a vehicle",
        "description": "Interprets the vehicle's latest Health Context through the reasoning agent and returns a grounded, persisted structured diagnosis.",
        "operationId": "ask_agent_api_v1_vehicles__vehicle_id__agent_query_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_id}/agent/events/critical": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Diagnose a critical telemetry rule event",
        "description": "Runs a critical-event diagnosis. Repeated events for the same vehicle and rule set within the cooldown window return the existing diagnosis instead of calling the LLM again.",
        "operationId": "diagnose_critical_event_api_v1_vehicles__vehicle_id__agent_events_critical_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriticalEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_id}/agent/dashboard": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Dashboard hook (no LLM call)",
        "description": "Reads-only snapshot for dashboards: the latest health context, the latest agent diagnosis and its age. Never invokes the LLM.",
        "operationId": "agent_dashboard_api_v1_vehicles__vehicle_id__agent_dashboard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_id}/agent/diagnoses": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "List agent diagnosis history",
        "description": "Paginated list of a vehicle's agent diagnoses, newest first.",
        "operationId": "list_agent_diagnoses_api_v1_vehicles__vehicle_id__agent_diagnoses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_AgentDiagnosisItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_id}/agent/diagnoses/latest": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Get the latest agent diagnosis",
        "description": "Returns the most recent agent diagnosis for the vehicle.",
        "operationId": "get_latest_agent_diagnosis_api_v1_vehicles__vehicle_id__agent_diagnoses_latest_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentQueryRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "AgentQueryRequest",
        "description": "A natural-language question about a vehicle."
      },
...
      "CriticalEventRequest": {
        "properties": {
          "rule_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Ids",
            "description": "Rule ids that triggered the critical event; derived from the context when omitted."
          }
        },
        "type": "object",
        "title": "CriticalEventRequest",
        "description": "A critical telemetry rule event to diagnose (rule ids optional)."
      },
      "DashboardResponse": {
        "properties": {
          "vehicle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vehicle Id"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "health_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HealthContextResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_diagnosis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagnosisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "diagnosis_age_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnosis Age Seconds"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "generated_at"
        ],
        "title": "DashboardResponse",
        "description": "Dashboard hook: current health context plus latest diagnosis (no LLM)."
      },
...
      "DiagnosisContent": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "possible_causes": {
            "items": {
              "$ref": "#/components/schemas/Hypothesis"
            },
            "type": "array",
            "title": "Possible Causes"
          },
          "recommended_actions": {
            "items": {
              "$ref": "#/components/schemas/RecommendedAction"
            },
            "type": "array",
            "title": "Recommended Actions"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "severity_analysis": {
            "$ref": "#/components/schemas/SeverityAnalysis"
          },
          "confidence_analysis": {
            "$ref": "#/components/schemas/ConfidenceAnalysis"
          },
          "context_note": {
            "type": "string",
            "title": "Context Note",
            "default": ""
          },
          "manufacturer_guidance": {
            "type": "string",
            "title": "Manufacturer Guidance",
            "default": ""
          },
          "cited_sources": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Cited Sources"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "manufacturer_evidence": {
            "items": {
              "$ref": "#/components/schemas/ManufacturerEvidence"
            },
            "type": "array",
            "title": "Manufacturer Evidence"
          }
        },
        "type": "object",
        "title": "DiagnosisContent",
        "description": "The interpreted diagnosis produced by the reasoning layer.\n\nContent is anchored to the Vehicle Health Context; anything that cannot be\ngrounded in that context is removed during validation.\n\nWhen the Manufacturer Guidance Tool supplied evidence, the model may add a\n``manufacturer_guidance`` note and ``cited_sources`` references. Only refs\nthat resolve to *actually retrieved* evidence survive validation; titles,\npage ranges and document ids are rebuilt app-side from the tool result."
      },
      "DiagnosisResponse": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "vehicle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Vehicle Id"
          },
          "trigger_type": {
            "type": "string",
            "title": "Trigger Type"
          },
          "user_query": {
            "type": "string",
            "title": "User Query",
            "default": ""
          },
          "diagnosis": {
            "$ref": "#/components/schemas/DiagnosisContent"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "critical"
            ],
            "title": "Severity"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "context_timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Timestamp"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "execution": {
            "$ref": "#/components/schemas/ExecutionMetadata"
          }
        },
        "type": "object",
        "required": [
          "vehicle_id",
          "trigger_type",
          "diagnosis",
          "severity",
          "confidence",
          "generated_at"
        ],
        "title": "DiagnosisResponse",
        "description": "The full agent response returned by the API."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Hypothesis": {
        "properties": {
          "cause": {
            "type": "string",
            "title": "Cause"
          },
          "likelihood": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Likelihood",
            "default": "medium"
          },
          "matching_evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matching Evidence"
          },
          "recommended_actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommended Actions"
          }
        },
        "type": "object",
        "required": [
          "cause"
        ],
        "title": "Hypothesis",
        "description": "A possible cause. Wording must treat it as a hypothesis, not a verdict."
      },
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LogoutRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "LogoutRequest"
      },
...
      "PaginatedResponse_AgentDiagnosisItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AgentDiagnosisItem"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "PaginatedResponse[AgentDiagnosisItem]"
      },
...
      "RefreshRequest": {
        "properties": {
          "refresh_token": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Refresh Token"
          }
        },
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "title": "RefreshRequest"
      },
      "RegisterRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Email used to sign in (case-insensitive)"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 8,
            "title": "Password",
            "description": "Password (bcrypt has 72-byte limit; 8+ chars required)"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "description": "Create a user account. Role and is_active are never client-controlled."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "refresh_token": {
            "type": "string",
            "title": "Refresh Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "user"
        ],
        "title": "TokenResponse"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}